+++ demisto_sdk/commands/content_graph/interface/neo4j/neo4j_graph.py
@@ -25,7 +25,6 @@ from demisto_sdk.commands.content_graph.interface.neo4j.queries.nodes import (_m
                                                                               remove_server_nodes)
 from demisto_sdk.commands.content_graph.interface.neo4j.queries.relationships import create_relationships
 from demisto_sdk.commands.content_graph.objects.base_content import BaseContent, ServerContent, content_type_to_model
-from demisto_sdk.commands.content_graph.objects.content_item import ContentItem
 from demisto_sdk.commands.content_graph.objects.integration import Integration
 from demisto_sdk.commands.content_graph.objects.pack import Pack
 from demisto_sdk.commands.content_graph.objects.relationship import RelationshipData
@@ -92,8 +91,6 @@ class Neo4jContentGraphInterface(ContentGraphInterface):
                 if not model:
                     raise NoModelException(f"No model for {content_type}")
                 obj = model.parse_obj(node)
-                if isinstance(obj, ContentItem):
-                    obj.fix_for_marketplace(marketplace)
                 Neo4jContentGraphInterface._id_to_obj[element_id] = obj
 
     def _add_relationships_to_objects(self, result: List[Neo4jResult]) -> List[BaseContent]:
@@ -315,7 +312,6 @@ class Neo4jContentGraphInterface(ContentGraphInterface):
     def create_pack_dependencies(self):
         with self.driver.session() as session:
             session.write_transaction(create_pack_dependencies)
-        super().create_pack_dependencies()
 
     def run_single_query(self, query: str, **kwargs) -> Any:
         with self.driver.session() as session: