+++ demisto_sdk/commands/content_graph/parsers/indicator_type.py
@@ -53,4 +53,4 @@ class IndicatorTypeParser(
             )
 
         if layout := self.json_data.get("layout"):
-            self.add_dependency_by_id(layout, ContentType.LAYOUT)
+            self.add_dependency_by_id(layout, ContentType.LAYOUT, is_mandatory=False)